drm/mipi-dsi: use correct return type for the DSC functions
From Dmitry Baryshkov
dda6efae161dc200b449d9099d0dfadfa75275c9 in linux-6.6.y/6.6.33
de1c705c50326acaceaf1f02bc5bf6f267c572bd in mainline linux

+++ sys/dev/pci/drm/include/drm/drm_mipi_dsi.h
@@ -70,8 +70,8 @@ ssize_t mipi_dsi_dcs_write(struct mipi_dsi_device *, u8, const void *, size_t);
 int mipi_dsi_dcs_nop(struct mipi_dsi_device *);
 int mipi_dsi_set_maximum_return_packet_size(struct mipi_dsi_device *, u16);
 bool mipi_dsi_packet_format_is_long(u8);
-ssize_t mipi_dsi_compression_mode(struct mipi_dsi_device *, bool);
-ssize_t mipi_dsi_picture_parameter_set(struct mipi_dsi_device *,
+int mipi_dsi_compression_mode(struct mipi_dsi_device *, bool);
+int mipi_dsi_picture_parameter_set(struct mipi_dsi_device *,
     const struct drm_dsc_picture_parameter_set *);
 
 static inline int
